fai_m10_v1/requirements.txt
# Production dependencies for secure financial RAG deployment
# Service: OPENAI (auto-detected from script)

# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Environment and configuration
python-dotenv==1.0.0

# AWS SDK (for VPC, KMS, Secrets Manager, CloudWatch, CloudTrail)
boto3==1.34.0
botocore==1.34.0

# OpenAI (primary RAG service)
openai==1.10.0

# Pinecone (vector database)
pinecone-client==3.0.0

# Database (PostgreSQL for user/role management)
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Caching (Redis)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0

# Security scanning
bandit==1.7.5

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.6

# Data processing
pandas==2.1.4
numpy==1.26.2

# Logging and monitoring
python-json-logger==2.0.7
